LG-10 befriend openGL WIP pt. 16

+++ src/source.h
@@ -10,13 +10,13 @@
 typedef short int direction_t;
 
 typedef struct lightpolygon { 
-    int x;
-    int y;
-    int red;
-    int green;
-    int blue;
-    int light_power;
-    int width;
+    int   x;
+    int   y;
+    float red;
+    float green;
+    float blue;
+    int   light_power;
+    int   width;
 } lightpolygon_t;
 
 typedef struct wobble {
